Parse fragment and attributes better. Detect unsoundness
- Now we can merge attributes same as we merge children. When the merged
result is a single node, we preserve <></> wrapper if edited version
uses so. Note that we only recgonize JSX fragment in attribute, and children.
Fragment at other places are treated as chrome of the Plasmic node that the Fragment wraps,
though it may be unsound if it wraps multiple Plasmic node. Such node will be treated
as deleted.

- value property of PlasmicSlot is handled specially - we always use the
edited version.
- slot args used in non argsRef node will also be renamed, such as this
one
  <div>{myFunction() && args.slot1}</div>

TODO: support renaming of non slot args
TODO: don't treat unsound node as deleted, but do a best effort merge,
i.e. treat the boundary at the JSX node.

Change-Id: Ia064da165d3de7e7e22133e10ca81d917dd55371