[cli] Improve auth token flow by avoiding copy/paste
This adds the concept of "init token", a random string generated by CLI
to poll the server for a personal access token ("auth token").

The workflow is the following:

1. User enters "plasmic init". CLI responds:

> What is your email address?
> Please log into this link:
https://studio.plasmic.app/auth/plasmic-init?token=<INIT_TOKEN>

<INIT_TOKEN> is generated using uuidv4.

Then, CLI starts a socket.io connection with
https://studio.plasmic.app/api/v1/init-token using email and
<INIT_TOKEN> as headers.

2. User logs into /auth/plasmic-init (layout in
https://studio.plasmic.app/projects/oYWs1jXLUht24zyQBdCd5F, may be
improved) and clicks "Authorize". Then the server will respond to the
CLI (via socketio, using email and init token as keys) with a newly
generated auth token.

3. CLI adds token to .plasmic.auth (as before).

Change-Id: I7140f56f2aaddd5e70e43d8732c5b438d6c4e6ba