[cli] Extend config to indicate codegen target (React, Next.js or Gatsby)
This improves platform detection to detect Gatsby, besides Next.js. It
also adds a prompt in `plasmic init` do choose which platform to
generate code for. After this patch, plasmic.json can contain
"platform": "react" | "nextjs" | "gatsby" (instead of only "react").

The idea of this is to be able to generate code that is specific to one
of these platforms. In particular, this is required to generate page
components links (see "Code generation target" in
https://coda.io/d/Plasmic-Eng-Wiki_dHQygjmQczq/Static-Jamstack-Site-Builder_suCMC)

Tested with projects using plain React, Next.js and Gatsby.

Change-Id: Id9fe39d32d947d662f5558136a52ebca2edf18fa