[cli] Assert all paths in plasmic.json are inside rootDir
Currently plasmic.json can have `srcDir` with any value, for example
`/usr/bin`. The same goes for `defaultPlasmicDir` and paths in general
in plasmic.json.

It could be "fine" for the user to use paths outside of root directory,
but it's dangerous from a security point of view when we are running
"plasmic sync" in a continuous deployment environment without an
isolated sandbox. Also that's probably not useful and may even indicate
an error in Plasmic configuration that leads to annoyingly writing into
weird directories; for example, a user could accidentally add or remove
a dot from the beginning default srcDir for Next.js, making it
"../components" or "/components".

This patch fixes that by creating a new function in file-utils named
`assertAllPathsInRootDir`. To avoid repeating the same logic of
`fixAllFilePaths` to grab all the file paths in a Plasmic configuration,
it also adds a function `getAllPaths` which is used by both functions.
That function grabs a bit more than fixAllFilePaths did before; it also
gets directories besides files (namely: defaultPlasmicDir, publicDir
and pagesDir).

Change-Id: Icc2e1da1e573186f708d77f3c70ebee98ef737e1