[token] Allow missing plasmic.json *and* skip auth
The plasmic.json config could be missing if CLI is run by PlasmicLoader,
but this shouldn't force an auth yet - we still need to check if the
tokens are in plasmic-loader.json.

Also, improve error messages since missing auth is tolerated now and you
could just be missing project API tokens.

https://app.clubhouse.io/plasmic/story/10408/move-away-from-secret-auth-tokens

Change-Id: Ifa293e2dca9a337d3eae715574471d25a51d6288