[plume] PLUME_STRICT_MODE can be toggled off for live mode
Some things, like select-option, when in non-strict mode, will not
hook in react-aria behavior, and instead just work like a normal
"dumb" component.

Also altered how aria-derived variants are explicitly toggled --
previously, we'd toggle isSelected if the option is selected, but
say nothing when the option is not selected. Now we explicitly
set isSelected=false when option is not selected.

Tweaked the API for defining a variant in the Plume hooks, so instead
of ["group", "variant"], can pass {group, variant}

Tweaked Select.children props to be ReactNode, and flatten it into
an array of elements, to support fragment children.

Change-Id: I6be6041ba3fbf61e206fc7201bdaa473c53a30c0