[overlay] Create triggered-overlay Plume component
It is a sub-component to Select and Menu for rendering the overlay
relative to some trigger.

Also a DropdownMenu component that can be used to compose a Menu
with any trigger the user wants, from React code (not used from
within Studio).

Also shared some logic around select and menu.

Change-Id: I1ccba29b2f45c9dfc1a7449bc1c67e77a555b77c