[loader-vue] Add loader vue compatible with vue2/3
Issue: https://app.clubhouse.io/plasmic/story/17167/pl2-does-not-work-with-vue-3

The main usage difference from before to now is that an user that is going to use Vue 2 has to install @vue/composition-api. Besides that the package can be used in both Vue 2 and Vue 3 in the same manner. One extra feature that is introduced for Vue 3 is `PlasmicVue` that allows the user to install in the app.

```
const app = createApp(App);
app.use(PlasmicVue);
app.mount("#app");
```

This way both PlasmicRootProvider and PlasmicComponent will be available to be used in any file, without the need to import every time from @plasmicapp/loader-angular.

Change-Id: Ic753151f7e44ad3e895d063e3cb831daed7d6e93