[host] <DisableWebpackHmr /> that swaps out window.EventSource
webpack hmr uses EventSource to listen to hot reloads, but that
results in a persistent connection from each window.  In Studio,
that means each artboard will be holding a persistent connection
when the app is custom-hosted by nextjs or gatsby servers in dev
mode.  Because browsers have a limit to how many connections can
be held at a time by domain, this means after X artboards, new
artboards will freeze and not load.

<DisableWebpackHmr /> will globally mutate window.EventSource
to avoid using EventSource for HMR.

Change-Id: I88931d3d72c9a2832939e53844f26ed33966c537