[state-management] Add registerInitFunc to $state
This patch adds a registerInitFunc method to support initializing data from $ctx.

Now, `specs` is a react state too because we need to update the changes from the `specs` passed to useDollarState.

To avoid infinite state update loop we only update the specs when one of the state initial value is updated. (the same idea from resetSpecs)

Change-Id: I4482a1d7e289e537167ec98eba127d9a66d1dc4d