[react-web] PlasmicSlot wraps content string DOM if <Trans/>
PlasmicSlot right now wraps slot content in
div.wab_slot-string-wrapper if the content is just a string. That's
because without an element, things like margin-left (if a child of
flex-gap) or even text-align:center does not get applied.

If the user turns on localization, then all strings are wrapped
in <Trans/>. These strings will also need the wrapper, as Trans
doesn't render any DOM elements. So we treat Trans the same way
we treat Fragments.

Change-Id: I827593fa2a5caf0aec5d86d1bd1804725de6f0ae