[host] Improve typechecking performance for host PropType
Following advice here: https://github.com/microsoft/TypeScript/wiki/Performance-Tracing

General technique is to switch to using interfaces instead
of intersections, which "flattens" the fields for each defined
interface. Then, we have one giant union of the flattened
interface types.  This is faster than using intersections
because tsc does not have to check against the same
intersected type over and over again for each member of the
giant union of types (I think?).

A few sacrifices were made - DefaultValueOrExpr and ControlTypeBase
are no more, as they are union types and can't be used
as interfaces to extend from. Instead, those fields are now
just baked into the base interfaces directly, which means we lose
the benefits of discriminated union. But these are minor
controls and a small price to pay.

Type-checking @plasmicpkgs/antd5 goes from ~30s to ~5s with this change.

Change-Id: I9239147b7845bb584d7abd5af2bce8cc5dd01e40