[data] Differentiate between null data key and dependent data key
This handles the case when one query depends on another query.

When our dataOp is null, it may be because we depend on something
that's not ready yet (so `$state.input.value` is null, for example),
or it may be because we depend on another data query (so
`$queries.query.data.hello` throws an undefined proxy error).
We need to properly differentiate between the two, because the right
thing to do is different -- in the former case, there's nothing
we can do except to try again when the component next renders, so
we mimic useSWR's behavior.  In the latter case, we want the
component to render again once the undefined proxy promise is resolved,
so we throw that promise.

It is important to throw that promise instead of doing nothing,
because the thrown promise may be the only thing that causes
React to try rendering this component again.  For example, suppose
we have one query fetching list of user IDs, second query fetching
a user based on one of the user IDs, and a text element that renders
the fetched user's name.  This happens:

* The second query depends on the first query, will encounter
  the undefined data promise when trying to access first query's
  result when building its dataOp.  It throws the promise.
* React will render the component again once that promise resolves
  (so the first query is done).
* React renders the element displaying user name, which will encounter
  undefined data promise as the second query is not ready yet.
  Promise is thrown.
* When the second query is resolved, React renders the component again
  and everything works!

If the second query DOESN'T throw the promise when using first query's
data, then we will never throw the first query's promise. Instead,
we proceed to render the element, and throws the second query's
promise instead. But the second query will never be resolved, because
it never got started running as it depended on first query's result.
But React will never know to try rendering this component again once
the first query is done, because we never threw the first query's
promise, so React doesn't know this component depends on it!

This CR modifies usePlasmicDataOp() to allow taking in a function that
resolves to a dataOp, instead of just a dataOp.  Then, when executing
the function, it can watch for PlasmicUndefinedDataError and act
accordingly.

This CR by itself is not sufficient; we'll also need to fix codegen
and canvas-rendering to pass in a dataOp function instead of a dataOp
directly.

Story: https://app.shortcut.com/plasmic/story/38205/issue-with-plasmicundefineddataerror

Change-Id: I704d2c55c3f8a2ad4929a2812f1485aacff7c744