[wab] Fix parts of outline missing
When the outline is closed, ListSpace sets the height as 0.
If an element is focused while the height is 0, useRevealOnFocus
will scroll the FixedSizeList. When the outline is opened,
the scroll persists, causing FixedSizeList to only render elements
after the scroll offset.

This may be a bug on FixedSizeList not adjusting scroll offset
automatically when the height changes.

To prevent this, don't render the FixedSizeList if height is 0.

Issue: https://linear.app/plasmic/issue/PLA-10365
Change-Id: I320ca9f47d4f6d0cb21b8d62ba07217ace380e26
GitOrigin-RevId: aa275ee6db998ac1712d41d7defbdf12fe742081