[rsbuild] Fix loading of css as string
rspack doesn't like it when we use !raw-loader! to load css files
as content; it always ends up as {}. So instead, our hack is to cp
the files we need in Makefile, as "txt" files, which raw-loader
happily loads.

For TinyMCE, could not get loading css as file references working
(file-loader), but loading and using as string seems to work just as
well there...

Change-Id: I7972941af9b60fc27e33a481bcdde2246ba73a5d
GitOrigin-RevId: 5d5fe991ecb11249b2bc8c36df81ef0399e897ed