fix: ensure we are using same host version (#1284)
When I was testing publishing test versions: #1221, I accidentally published @plasmicapp/host@2.0.0-test.0 with the latest tag for a short duration. I have since removed this version and restored the latest to 1.0.222. Unfortunately, our upgrade-internal process picked up the v2 version during the duration. It's been like this for 2 weeks, so it also means that the v2 is version is compatible (which is expected). However, I'm reverting back since we should be using a consistent host version.

GitOrigin-RevId: 69f5f911ae9a8ba45ceb1c0dd55c8fe576f97c99