Implement nested ref resolution (#2018)
This commit implements nested ref resolution. Nested ref fields can be requested in `q.fields`. Example inputs are "ref.name" or "nestedList.nestedRef.name".

The algorithm for resolving nested refs has 3 phases:
1. makeSelectionTree - Build tree of selected fields and refs to resolve based on `q.fields` and schemas. This tree will be used in later phases to know where to find and replace refs.
2. resolveNested - Efficiently collect refs to be resolved and load rows batched per table into a row cache.
3. projectCmsData - Replace all refs with rows in the cache.

A lot of the complexity is in makeSelectionTree. The challenge is parsing the field paths in `q.fields` and organizing them in a sane manner to build the tree. The key here is toPartitionedPath, which partitions the field path into an array in the format [...RefPartition, TerminalPartition]. RefPartition can either be a ref field (simple case) or a ref field inside a list/object field (complex case). TerminalPartition is either a "*" or any field.
Examples that illustrate toPartitionedPath:
"ref.*" => [[ref], *]
"nestedList.nestedRef.name" => [[nestedList, nestedRef], name]

GitOrigin-RevId: f39481a192f095f7b2929d84af542ac4af0bcbb3